bpf: add a new binary_path struct to execve_map_value
This new struct will store information about the binary path from the
moment we persist the execve information in the execve_map. We tried to
reduce at maximum the size of what we store, ending up with 256 bytes
instead of the theoretical maximum MAX_PATH 4096 bytes (+ metadata).

This will be useful when doing the matchBinary at a later stage and
retrieving the information about the process from the execve_map.

Signed-off-by: Mahe Tardy <mahe.tardy@gmail.com>