pkg/sensors/exec: add the binary_path struct to userspace
A new binary_path struct was added BPF side to store a part of the
binary path inside the execve_map values (to do comparison at a later
stage). This commits introduces the changes on the proc reader part,
which scans /proc at startup and initialize/fill execve_map with
information of processes that were started before tetragon.

Also it moves the part that is trimming the p.args if the size of the
process information would not fit in the allocated buffer. We were
previously doing it in the part that parses /proc, which was too early
because in the case of execve_map initialization, it's not needed (and
we now need at least 255 bytes of the binary path guaranted), while it's
needed for pushing the execve event, where it was moved.

We also needed the 'exe' value at execve_map initialization, which was
already merged with 'cmdline' early at /proc parsing since it was not
necessary previously. Now we merge 'exe' and 'cmdline' on demand at a
later stage (again when pushing the execve event).

Signed-off-by: Mahe Tardy <mahe.tardy@gmail.com>